+++ internal/rest-client/space.go
@@ -2,8 +2,8 @@ package client
 
 type Space struct {
 	DataReduction float64 `json:"data_reduction"`
-	Snapshots     float64 `json:"snapshots"`
-	TotalPhysical float64 `json:"total_physical"`
-	Unique        float64 `json:"unique"`
-	Virtual       float64 `json:"virtual"`
+	Snapshots     int64   `json:"snapshots"`
+	TotalPhysical int64   `json:"total_physical"`
+	Unique        int64   `json:"unique"`
+	Virtual       int64   `json:"virtual"`
 }
